annonce et pagination

+++ src/main/java/com/commercial/commerce/sale/controller/CategoryController.java
@@ -39,7 +39,7 @@ public class CategoryController extends Controller {
     }
 
     @GetMapping("/actu/categories/pagination")
-    public ResponseEntity<ApiResponse<Integer>> getPagination(
+    public ResponseEntity<ApiResponse<Long>> getPagination(
             @RequestParam(name = "limit", defaultValue = "5") int limit) {
         try {
