+++ src/main/java/com/commercial/commerce/sale/controller/TransactionController.java
@@ -122,7 +122,7 @@ public class TransactionController extends Controller {
     }
 
     @GetMapping("/transactions/pagination")
-    public ResponseEntity<ApiResponse<Integer>> getPagination(
+    public ResponseEntity<ApiResponse<Long>> getPagination(
             @RequestParam(name = "limit", defaultValue = "5") int limit) {
         try {
