+++ src/main/java/com/commercial/commerce/sale/controller/TypeController.java
@@ -114,7 +114,7 @@ public class TypeController extends Controller {
     }
 
     @GetMapping("/actu/types/pagination")
-    public ResponseEntity<ApiResponse<Integer>> getPagination(
+    public ResponseEntity<ApiResponse<Long>> getPagination(
             @RequestParam(name = "limit", defaultValue = "5") int limit) {
         try {
